move the 'use client' to an client.js to use the metadata correctly
