+++ src/app/(dashboard)/app/tokens/page.js
@@ -1,8 +1,4 @@
-'use client'
-import { useState } from 'react'
-
 import TokenList from '@/components/Tokens/TokenList'
-import 'react-toastify/dist/ReactToastify.css'
 
 export const metadata = {
 	title: 'API Tokens - ForVoyez',
@@ -14,8 +10,6 @@ export const metadata = {
 }
 
 export default function TokenPage() {
-	const [tokens, setTokens] = useState([])
-
 	return (
 		<div className={'prose mx-auto max-w-5xl flex-auto'}>
 			<h1 className="text-xl font-bold text-slate-800">Manage Tokens</h1>
