+++ src/app/(dashboard)/app/usage/page.js
@@ -1,7 +1,3 @@
-'use client'
-
-import React from 'react'
-
 import { UsageChartComponent } from '@/components/usage/usageChart.component'
 
 export const metadata = {
@@ -17,7 +13,6 @@ export default function PlansPage() {
 	return (
 		<>
 			<h1 className="my-6 text-center text-2xl font-bold">API Usage</h1>
-
 			<UsageChartComponent />
 		</>
 	)
